General fixes to seared furnace
