+++ src/main/java/slimeknights/tconstruct/smeltery/client/GuiSmeltery.java
@@ -1,60 +1,44 @@
 package slimeknights.tconstruct.smeltery.client;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
-
 import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.TextFormatting;
-import net.minecraftforge.fluids.Fluid;
 import net.minecraftforge.fluids.FluidStack;
 
 import java.io.IOException;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
-
 import slimeknights.mantle.client.gui.GuiElement;
-import slimeknights.mantle.client.gui.GuiMultiModule;
 import slimeknights.tconstruct.common.TinkerNetwork;
-import slimeknights.tconstruct.library.TinkerRegistry;
 import slimeknights.tconstruct.library.Util;
 import slimeknights.tconstruct.library.client.RenderUtil;
-import slimeknights.tconstruct.library.smeltery.CastingRecipe;
-import slimeknights.tconstruct.library.smeltery.ICastingRecipe;
 import slimeknights.tconstruct.library.smeltery.SmelteryTank;
-import slimeknights.tconstruct.smeltery.TinkerSmeltery;
-import slimeknights.tconstruct.smeltery.client.module.GuiSmelterySideinventory;
+import slimeknights.tconstruct.smeltery.client.module.GuiSmelterySideInventory;
 import slimeknights.tconstruct.smeltery.inventory.ContainerSmeltery;
 import slimeknights.tconstruct.smeltery.network.SmelteryFluidClicked;
-import slimeknights.tconstruct.smeltery.tileentity.TileHeatingStructureFuelTank;
 import slimeknights.tconstruct.smeltery.tileentity.TileSmeltery;
 import slimeknights.tconstruct.tools.inventory.ContainerSideInventory;
 
-public class GuiSmeltery extends GuiMultiModule {
+public class GuiSmeltery extends GuiHeatingStructureFuelTank {
 
   public static final ResourceLocation BACKGROUND = Util.getResource("textures/gui/smeltery.png");
 
   protected GuiElement scala = new GuiElement(176, 76, 52, 52, 256, 256);
 
-  protected final GuiSmelterySideinventory sideinventory;
+  protected final GuiSmelterySideInventory sideinventory;
   protected final TileSmeltery smeltery;
-
-  private TileHeatingStructureFuelTank.FuelInfo fuelInfo;
 
   public GuiSmeltery(ContainerSmeltery container, TileSmeltery smeltery) {
     super(container);
 
     this.smeltery = smeltery;
 
-    sideinventory = new GuiSmelterySideinventory(this, container.getSubContainer(ContainerSideInventory.class),
+    sideinventory = new GuiSmelterySideInventory(this, container.getSubContainer(ContainerSideInventory.class),
                                                  smeltery, smeltery.getSizeInventory(), container.calcColumns());
     addModule(sideinventory);
   }
 
+  // this is the same for both structures, but the superclass does not have (nor need) access to the side inventory
   @Override
   public void updateScreen() {
     super.updateScreen();
@@ -96,25 +80,14 @@ public class GuiSmeltery extends GuiMultiModule {
       }
       else {
         text.add(TextFormatting.WHITE + hovered.getLocalizedName());
-        liquidToString(hovered, text);
+        RenderUtil.liquidToString(hovered, text);
       }
 
       this.drawHoveringText(text, mouseX, mouseY);
     }
     // Fuel
     else if(71 <= mouseX && mouseX < 83 && 16 <= mouseY && mouseY < 68) {
-      List<String> text = Lists.newArrayList();
-      FluidStack fuel = fuelInfo.fluid;
-      text.add(TextFormatting.WHITE + Util.translate("gui.smeltery.fuel"));
-      if(fuel != null) {
-        text.add(fuel.getLocalizedName());
-        liquidToString(fuel, text);
-      }
-      else {
-        text.add(Util.translate("gui.smeltery.fuel.empty"));
-      }
-      text.add(Util.translateFormatted("gui.smeltery.fuel.heat", fuelInfo.heat));
-      this.drawHoveringText(text, mouseX, mouseY);
+      drawFuelTooltip(mouseX, mouseY);
     }
   }
 
@@ -144,15 +117,7 @@ public class GuiSmeltery extends GuiMultiModule {
 
     // update fuel info
     fuelInfo = smeltery.getFuelDisplay();
-
-    if(fuelInfo.fluid != null && fuelInfo.fluid.amount > 0) {
-      int x = 71 + cornerX;
-      int y = 16 + cornerY + 52;
-      int w = 12;
-      int h = (int) (52f * fuelInfo.fluid.amount / fuelInfo.maxCap);
-
-      RenderUtil.renderTiledFluid(x, y - h, w, h, this.zLevel, fuelInfo.fluid);
-    }
+    drawFuel(71, 16, 12, 52);
   }
 
   @Override
@@ -197,97 +162,4 @@ public class GuiSmeltery extends GuiMultiModule {
     return SmelteryRenderer.calcLiquidHeights(liquids, capacity, scala.h, 3);
   }
 
-  /* Fluid amount displays */
-  private static Map<Fluid, List<FluidGuiEntry>> fluidGui = Maps.newHashMap();
-
-  public void liquidToString(FluidStack fluid, List<String> text) {
-    int amount = fluid.amount;
-    if(!Util.isShiftKeyDown()) {
-      List<FluidGuiEntry> entries = fluidGui.get(fluid.getFluid());
-      if(entries == null) {
-        entries = calcFluidGuiEntries(fluid.getFluid());
-        fluidGui.put(fluid.getFluid(), entries);
-      }
-
-      for(FluidGuiEntry entry : entries) {
-        amount = calcLiquidText(amount, entry.amount, entry.getText(), text);
-      }
-    }
-
-    // standard display: bucket amounts
-    // we go up to kiloBuckets because we can
-    amount = calcLiquidText(amount, 1000000, Util.translate("gui.smeltery.liquid.kilobucket"), text);
-    amount = calcLiquidText(amount, 1000, Util.translate("gui.smeltery.liquid.bucket"), text);
-    calcLiquidText(amount, 1, Util.translate("gui.smeltery.liquid.millibucket"), text);
-  }
-
-  private List<FluidGuiEntry> calcFluidGuiEntries(Fluid fluid) {
-    List<FluidGuiEntry> list = Lists.newArrayList();
-
-    // go through all casting recipes for the fluids and check for known "units" like blocks, ingots,...
-    for(ICastingRecipe irecipe : TinkerRegistry.getAllBasinCastingRecipes()) {
-      if(irecipe instanceof CastingRecipe) {
-        CastingRecipe recipe = (CastingRecipe) irecipe;
-        // search for a block recipe
-        if(recipe.getFluid().getFluid() == fluid && recipe.cast == null) {
-          // it's a block that is cast solely from the material, using no cast, therefore it's a block made out of the material
-          list.add(new FluidGuiEntry(recipe.getFluid().amount, "gui.smeltery.liquid.block"));
-        }
-      }
-    }
-    // table casting
-    for(ICastingRecipe irecipe : TinkerRegistry.getAllTableCastingRecipes()) {
-      if(irecipe instanceof CastingRecipe) {
-        CastingRecipe recipe = (CastingRecipe) irecipe;
-        if(recipe.getFluid().getFluid() == fluid && recipe.cast != null) {
-          // nugget
-          if(recipe.cast.matches(new ItemStack[]{TinkerSmeltery.castNugget}) != null) {
-            list.add(new FluidGuiEntry(recipe.getFluid().amount, "gui.smeltery.liquid.nugget"));
-          }
-          // ingot
-          if(recipe.cast.matches(new ItemStack[]{TinkerSmeltery.castIngot}) != null) {
-            list.add(new FluidGuiEntry(recipe.getFluid().amount, "gui.smeltery.liquid.ingot"));
-          }
-          // gem
-          if(recipe.cast.matches(new ItemStack[]{TinkerSmeltery.castGem}) != null) {
-            list.add(new FluidGuiEntry(recipe.getFluid().amount, "gui.smeltery.liquid.gem"));
-          }
-        }
-      }
-    }
-
-    // sort by amount descending because the order in which they're accessed is important since it changes the remaining value during processing
-    Collections.sort(list, new Comparator<FluidGuiEntry>() {
-      @Override
-      public int compare(FluidGuiEntry o1, FluidGuiEntry o2) {
-        return o2.amount - o1.amount;
-      }
-    });
-
-    return ImmutableList.copyOf(list);
-  }
-
-  private int calcLiquidText(int amount, int divider, String unit, List<String> text) {
-    int full = amount / divider;
-    if(full > 0) {
-      text.add(String.format("%d %s%s", full, TextFormatting.GRAY, unit));
-    }
-
-    return amount % divider;
-  }
-
-  private static class FluidGuiEntry {
-
-    public final int amount;
-    public final String unlocName;
-
-    private FluidGuiEntry(int amount, String unlocName) {
-      this.amount = amount;
-      this.unlocName = unlocName;
-    }
-
-    public String getText() {
-      return Util.translate(unlocName);
-    }
-  }
 }
